refactor(python): merge nuri.core._core into nuri.core
Make nuri.core itself the compiled extension (like nuri.fmt) rather
than a package re-exporting a private nuri.core._core submodule, so the
core classes report __module__ == "nuri.core" and the private path no
longer leaks into signatures, stubs, or imports. Fold the former
geometry extension in as a def_submodule, preserving the
nuri.core.geometry import path, and relocate the pure-Python masquerade
helper to nuri._masquerade_support.
