+++ cmake/NuriKitPythonUtils.cmake
@@ -22,7 +22,7 @@ function(nuri_python_generate_stubs module output dependency)
     "${PYBIND11_STUBGEN}"
     -o "${CMAKE_CURRENT_LIST_DIR}"
     --exit-code
-    --enum-class-locations .*:nuri.core._core
+    --enum-class-locations .*:nuri.core
     --numpy-array-remove-parameters
     ${ARGN}
     "${module}"
