--- python/src/nuri/core/__init__.py
@@ -1,37 +1,0 @@
-#
-# Project NuriKit - Copyright 2023 SNU Compbio Lab.
-# SPDX-License-Identifier: Apache-2.0
-#
-"""
-The core module of NuriKit.
-
-This module contains the core classes of NuriKit. The core module is not
-very useful by itself, but is a dependency of many other modules. Chemical
-data structures, such as elements, isotopes, and molecules, and also the
-graph structure and algorithms, are defined in this module."""
-
-__all__ = [
-    "Atom",
-    "AtomData",
-    "Bond",
-    "BondConfig",
-    "BondData",
-    "BondOrder",
-    "Chirality",
-    "Element",
-    "Hyb",
-    "Isotope",
-    "Molecule",
-    "Mutator",
-    "Neighbor",
-    "PeriodicTable",
-    "SubAtom",
-    "SubBond",
-    "SubNeighbor",
-    "Substructure",
-    "SubstructureCategory",
-    "SubstructureContainer",
-    "seed_thread",
-]
-
-from ._core import *
